move RuleBuilder
